requirements.txt
flask
flask-cors
scikit-learn
nltk
spacy
torch
transformers
sentence-transformers
pdfplumber
python-docx
web3
ipfshttpclient
requests
pandas
wikipedia-api
arxiv
cryptography
